requirements.txt
# Framework Web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Validación y schemas
pydantic==2.5.0
pydantic[email]==2.5.0

# Seguridad y autenticación
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.1

# Pagos con Stripe
stripe==7.5.0

# Variables de entorno
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# MinIO (almacenamiento de archivos)
minio==7.2.0

# Utilidades
python-dateutil==2.8.2